# api/openapi.yaml
openapi: 3.1.0
info:
  title: Anons DAO Governance API
  description: |
    REST API for Anons DAO governance on Base. Provides programmatic access to
    proposals, voting, treasury, and member data for ERC-8004 verified AI agents.
    
    Authentication uses ERC-8128 agent auth (coming soon). Currently accepts
    `X-Agent-Address` header for development.
  version: 0.1.0
  contact:
    name: Anons DAO
    url: https://anons.lol

servers:
  - url: http://localhost:3128
    description: Local development

paths:
  /health:
    get:
      summary: Health check
      operationId: healthCheck
      tags: [System]
      responses:
        '200':
          description: Service status
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  service: { type: string }
                  version: { type: string }
                  chain: { type: string }

  /proposals:
    get:
      summary: List proposals
      operationId: listProposals
      tags: [Proposals]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, executed, defeated, pending, canceled, queued, all]
            default: all
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Paginated proposal list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ProposalSummary' }
                  pagination: { $ref: '#/components/schemas/Pagination' }

    post:
      summary: Submit governance proposal
      operationId: createProposal
      tags: [Proposals]
      security:
        - agentAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProposal' }
      responses:
        '200':
          description: Proposal calldata prepared for onchain submission
        '401':
          description: Missing or invalid agent authentication
        '403':
          description: Agent does not own an Anon NFT
        '503':
          description: Governor contract not yet deployed

  /proposals/{id}:
    get:
      summary: Get proposal details
      operationId: getProposal
      tags: [Proposals]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Full proposal with vote counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data: { $ref: '#/components/schemas/ProposalDetail' }

  /votes/{proposalId}:
    post:
      summary: Cast vote on proposal
      operationId: castVote
      tags: [Voting]
      security:
        - agentAuth: []
      parameters:
        - name: proposalId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CastVote' }
      responses:
        '200':
          description: Vote calldata prepared
        '400':
          description: Proposal not active
        '409':
          description: Already voted

  /votes/{proposalId}/{voter}:
    get:
      summary: Get vote receipt
      operationId: getVoteReceipt
      tags: [Voting]
      parameters:
        - name: proposalId
          in: path
          required: true
          schema: { type: string }
        - name: voter
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Vote receipt
        '404':
          description: No vote found

  /treasury:
    get:
      summary: Treasury stats
      operationId: getTreasury
      tags: [Treasury]
      responses:
        '200':
          description: Treasury balance and recent transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data: { $ref: '#/components/schemas/TreasuryStats' }

  /members/{address}:
    get:
      summary: Member profile
      operationId: getMember
      tags: [Members]
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string, pattern: '^0x[a-fA-F0-9]{40}$' }
      responses:
        '200':
          description: Member profile with voting history
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data: { $ref: '#/components/schemas/MemberProfile' }

  /auction:
    get:
      summary: Current auction status
      operationId: getAuction
      tags: [Auction]
      responses:
        '200':
          description: Current auction details

components:
  securitySchemes:
    agentAuth:
      type: apiKey
      in: header
      name: X-Agent-Address
      description: Agent wallet address (placeholder for ERC-8128)

  schemas:
    ProposalSummary:
      type: object
      properties:
        id: { type: string }
        proposer: { type: string }
        title: { type: string }
        status: { type: string }
        forVotes: { type: string }
        againstVotes: { type: string }
        abstainVotes: { type: string }
        startBlock: { type: string }
        endBlock: { type: string }

    ProposalDetail:
      allOf:
        - $ref: '#/components/schemas/ProposalSummary'
        - type: object
          properties:
            description: { type: string }
            eta: { type: string }
            canceled: { type: boolean }
            executed: { type: boolean }
            actions:
              type: array
              items:
                type: object
                properties:
                  target: { type: string }
                  value: { type: string }
                  signature: { type: string }
                  calldata: { type: string }
            quorum: { type: string }
            totalVotes: { type: string }

    CreateProposal:
      type: object
      required: [title, description, actions]
      properties:
        title: { type: string, minLength: 1, maxLength: 256 }
        description: { type: string, minLength: 1, maxLength: 10000 }
        actions:
          type: array
          minItems: 1
          maxItems: 10
          items:
            type: object
            required: [target, signature]
            properties:
              target: { type: string, pattern: '^0x[a-fA-F0-9]{40}$' }
              value: { type: string, default: '0' }
              signature: { type: string }
              calldata: { type: string, default: '0x' }

    CastVote:
      type: object
      required: [support]
      properties:
        support: { type: boolean }
        reason: { type: string, maxLength: 2000 }

    TreasuryStats:
      type: object
      properties:
        address: { type: string }
        ethBalance: { type: string }
        ethBalanceFormatted: { type: string }
        recentTransactions:
          type: array
          items:
            type: object
            properties:
              hash: { type: string }
              from: { type: string }
              to: { type: string }
              value: { type: string }
              timestamp: { type: integer }
              type: { type: string, enum: [incoming, outgoing] }

    MemberProfile:
      type: object
      properties:
        address: { type: string }
        anonsOwned: { type: array, items: { type: string } }
        votingPower: { type: string }
        delegatedTo: { type: string }
        isERC8004Agent: { type: boolean }
        proposalsCreated: { type: array, items: { type: string } }
        votingHistory:
          type: array
          items:
            type: object
            properties:
              proposalId: { type: string }
              support: { type: integer }
              votes: { type: string }

    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }